Improved linux compatibility

+++ SF_Cairo/RS.c
@@ -472,10 +472,10 @@ void SF_iteration(cairo_t *cr)
 	// The game should always wait 50ms between frames, so sleep until the loop
 	// body takes 50ms
 	// is this valid c?
-		elapsed_time=((clock()-loop_start_time)/(double)CLOCKS_PER_SEC)*1000.0;
+		elapsed_time=((clock()-loop_start_time)/(long double)CLOCKS_PER_SEC)*1000.0;
     if(elapsed_time < SF_DELAY)
-		{			
-//				printf("Sleeping for %Lf \n", elapsed_time);
+		{
+//				printf("Sleeping for %Lf \n", SF_DELAY-elapsed_time);
         usleep(SF_DELAY-elapsed_time);  /* wait up to 50 milliseconds */
 		}
 	} else Effect_Flag=OFF;  /* no delay necessary */
